+++ packages/rpc-error/CHANGELOG.md
@@ -1,3 +1,11 @@
+## v0.2.1 (2018-07-31)
+
+Fix `toObject()` method.
+
+## v0.2.0 (2018-07-31)
+
+Add `toObject()` method.
+
 ## v0.1.1 (2018-07-09)
 
 Export as CommonJS and ESM.